Remove unnecessary add_compile_defs abstraction
The two CMake functions `add_compile_defs` and `target_compile_defs`
were unnecessary. The just wrapped the CMake built-in commands
`add_compile_definitions` and `target_compile_definitions` with an extra
parameter, languages, which was then added in a generator expression for
every other argument. This is pointless, however, as CMake natively
understands the difference between a C library (with .c files) and a C++
library (with .cpp files). CMake will not compile a C library with a C++
compiler. I found no explanation in the history as to why this
abstraction was added, and nothing broke when it was removed.